Add a datasource config flag for making host_type mandatory
If 'require_host_type' is present in the datasource config and truthy,
a RDBSHubExecuteError exception will be raised if host_type is not
defined, rather than the normal fall-back to a type of 'master_host'.
This makes it easier to catch cases where 'host_type' was omitted from
the procs file, and queries would otherwise have run against the master
host unnecessarily.

Fixes #21.