Make run-tests.py exit(1) if tests failed
So that Travis will actually turn red if tests fail.

run-tests.py is overengineed for running the two test files there are
in the repo at the moment (and unnecessarily re-invents the wheel),
but for now it will do to at least get some tests running on Travis.
Later I'll probably convert the tests to use py.test to make it
easier to import Datasource into the Treeherder repo. This will also
fix the current bug where the test runner fails if |setup.py install|
has been run, due to path issues with the build directory it creates.