Simplify DataHub.get() and make it Python 3 compatible
`os.path.walk()` has been removed in Python 3, so needs to be replaced.
However there really is no need for it, since other than for providing a
list of available hubs in error messages etc, we don't need to know the
available hubs in advance. It's clearer to just use a try-except around
`importlib` (which is recommended over `imp` for Python 2.7+) instead.